+++ src/main/java/org/prgms/springbootbasic/repository/CustomerFileRepository.java
@@ -1,6 +1,6 @@
 package org.prgms.springbootbasic.repository;
 
-import org.prgms.springbootbasic.domain.customer.CustomerCsvFileManager;
+import org.prgms.springbootbasic.common.file.CustomerCsvFileManager;
 import org.prgms.springbootbasic.domain.customer.Customer;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
