+++ src/main/java/org/prgms/springbootbasic/repository/VoucherFileRepository.java
@@ -1,7 +1,7 @@
 package org.prgms.springbootbasic.repository;
 
 import lombok.extern.slf4j.Slf4j;
-import org.prgms.springbootbasic.domain.voucher.VoucherCsvFileManager;
+import org.prgms.springbootbasic.common.file.VoucherCsvFileManager;
 import org.prgms.springbootbasic.domain.voucher.VoucherPolicy;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
